+++ voxygen/src/hud/minimap.rs
@@ -1,6 +1,6 @@
 use super::{
     img_ids::{Imgs, ImgsRot},
-    Show, HP_COLOR, TEXT_COLOR, UI_HIGHLIGHT_0, UI_MAIN,
+    Show, TEXT_COLOR, UI_HIGHLIGHT_0, UI_MAIN,
 };
 use crate::ui::{fonts::ConrodVoxygenFonts, img_ids};
 use client::{self, Client};
@@ -8,10 +8,9 @@ use common::{comp, terrain::TerrainChunkSize, vol::RectVolSize};
 use conrod_core::{
     color, position,
     widget::{self, Button, Image, Rectangle, Text},
-    widget_ids, Color, Colorable, Positionable, Sizeable, Widget, WidgetCommon,
+    widget_ids, Colorable, Positionable, Sizeable, Widget, WidgetCommon,
 };
 use specs::WorldExt;
-use std::time::{Duration, Instant};
 use vek::*;
 
 widget_ids! {
@@ -68,8 +67,6 @@ impl<'a> MiniMap<'a> {
 pub struct State {
     ids: Ids,
 
-    last_region_name: Option<String>,
-    last_update: Instant,
     zoom: f64,
 }
 
@@ -86,8 +83,6 @@ impl<'a> Widget for MiniMap<'a> {
         State {
             ids: Ids::new(id_gen),
 
-            last_region_name: None,
-            last_update: Instant::now(),
             zoom: {
                 let min_world_dim = self.world_map.1.reduce_partial_min() as f64;
                 min_world_dim.min(
@@ -248,57 +243,6 @@ impl<'a> Widget for MiniMap<'a> {
         .was_clicked()
         {
             return Some(Event::Toggle);
-        }
-
-        // Display zone name on entry
-
-        const FADE_IN: f32 = 0.5;
-        const FADE_HOLD: f32 = 1.0;
-        const FADE_OUT: f32 = 3.0;
-        match self.client.current_chunk() {
-            Some(chunk) => {
-                let current = chunk.meta().name();
-                // Check if no other popup is displayed and a new one is needed
-                if state.last_update.elapsed()
-                    > Duration::from_secs_f32(FADE_IN + FADE_HOLD + FADE_OUT)
-                    && state
-                        .last_region_name
-                        .as_ref()
-                        .map(|l| l != current)
-                        .unwrap_or(true)
-                {
-                    // Update last_region
-                    state.update(|s| s.last_region_name = Some(current.to_owned()));
-                    state.update(|s| s.last_update = Instant::now());
-                }
-
-                let seconds = state.last_update.elapsed().as_secs_f32();
-                let fade = if seconds < FADE_IN {
-                    seconds / FADE_IN
-                } else if seconds < FADE_IN + FADE_HOLD {
-                    1.0
-                } else {
-                    (1.0 - (seconds - FADE_IN - FADE_HOLD) / FADE_OUT).max(0.0)
-                };
-                // Region Name
-                Text::new(state.last_region_name.as_ref().unwrap_or(&"".to_owned()))
-                    .mid_top_with_margin_on(ui.window, 200.0)
-                    .font_size(self.fonts.alkhemi.scale(70))
-                    .font_id(self.fonts.alkhemi.conrod_id)
-                    .color(Color::Rgba(0.0, 0.0, 0.0, fade))
-                    .set(state.ids.zone_display_bg, ui);
-                Text::new(state.last_region_name.as_ref().unwrap_or(&"".to_owned()))
-                    .top_left_with_margins_on(state.ids.zone_display_bg, -2.5, -2.5)
-                    .font_size(self.fonts.alkhemi.scale(70))
-                    .font_id(self.fonts.alkhemi.conrod_id)
-                    .color(Color::Rgba(1.0, 1.0, 1.0, fade))
-                    .set(state.ids.zone_display, ui);
-            },
-            None => Text::new(" ")
-                .middle_of(ui.window)
-                .font_size(self.fonts.alkhemi.scale(14))
-                .color(HP_COLOR)
-                .set(state.ids.zone_display, ui),
         }
 
         // TODO: Subregion name display